Refactor Orders packages
