+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/ui/mystore/MyStoreScreen.kt
@@ -5,8 +5,8 @@ import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import com.google.android.horologist.compose.pager.PagerScreen
-import com.woocommerce.android.ui.orders.OrdersListScreen
-import com.woocommerce.android.ui.orders.OrdersListViewModel
+import com.woocommerce.android.ui.orders.list.OrdersListScreen
+import com.woocommerce.android.ui.orders.list.OrdersListViewModel
 import com.woocommerce.android.ui.stats.StoreStatsScreen
 import com.woocommerce.android.ui.stats.StoreStatsViewModel
